JAVA/ChatGPT/Strategy/round-2/source/src/src/main/java/com/example/src/strategies/JWTAuthStrategy.java
package com.example.src.strategies;


import io.jsonwebtoken.Jwts;
import io.jsonwebtoken.SignatureAlgorithm;
import javax.servlet.http.HttpServletRequest;

public class JWTAuthStrategy implements Authenticator {
    private static final String SECRET_KEY = "secret";

    @Override
    public boolean authenticate(HttpServletRequest request) {
        String authHeader = request.getHeader("Authorization");
        if (authHeader != null && authHeader.startsWith("Bearer ")) {
            String token = authHeader.substring(7);
            try {
                Jwts.parser().setSigningKey(SECRET_KEY).parseClaimsJws(token);
                return true;
            } catch (Exception e) {
                return false;
            }
        }
        return false;
    }
}